fix: move util for telegram to api directory
